% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_url.r
\name{get_cdc_url}
\alias{get_cdc_url}
\title{Get CDC data URL for a given section and year}
\usage{
get_cdc_url(sec, yr)
}
\arguments{
\item{sec}{A string identifying the data section, e.g. \code{"mortality_multiple"}.}

\item{yr}{An integer or numeric year, e.g. \code{1969}.}
}
\value{
A character string containing the URL to the zip file on the CDC FTP server.
}
\description{
Looks up the US data download URL from the internal CDC link table for a
given section and year combination.
}
\examples{
\dontrun{
get_cdc_url("mortality_multiple", 1969)
}

}
