% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_mult_mort_user_guide.r
\name{scrape_mult_mort_user_guide}
\alias{scrape_mult_mort_user_guide}
\title{Scrape Mortality Multiple Cause-of-Death user guide links}
\usage{
scrape_mult_mort_user_guide(url)
}
\arguments{
\item{url}{URL of the CDC mortality documentation page.}
}
\value{
A tibble with columns:
\describe{
\item{section}{Always "mortality_multiple"}
\item{subsection}{Always "User Guide"}
\item{link_text}{Text of the download link}
\item{year}{Four-digit year extracted from link text, filled down for sub-items}
\item{file_size}{File size string, if present}
\item{url}{Absolute URL to the file}
\item{file_type}{File extension}
\item{file_size_mb}{File size converted to megabytes}
}
}
\description{
Extracts downloadable file links from the CDC Mortality Multiple
Cause-of-Death documentation page and returns a tidy tibble with
metadata about each file.
}
\details{
This function is designed for the CDC Mortality Multiple Cause-of-Death
documentation page at:
\url{https://www.cdc.gov/nchs/nvss/mortality_public_use_data.htm}

Note: 1997 and 1998 entries link to separate HTML pages containing many
PDFs and are not scraped by this function as of 2/17/2026.
}
